requirements.txt
smbus2
Jetson.GPIO
ADS1x15-ADC